[IMP] purchase_order_kmitl: merge department and OU modules
