+++ account_asset_purchase/__manifest__.py
@@ -6,7 +6,7 @@
     "author": "Aginix Technologies",
     "website": "https://github.com/aginix/kmitl",
     "category": "KMITL",
-    'depends': ['account_asset_number_kmitl', 'purchase_order_kmitl', 'l10n_th_gov_gpsc', 'account_asset_usability_kmitl', 'purchase_order_department', 'hr_department_operating_unit'],
+    'depends': ['account_asset_number_kmitl', 'purchase_order_kmitl', 'l10n_th_gov_gpsc', 'account_asset_usability_kmitl'],
     "data": [
         "security/ir.model.access.csv",
         "views/account_asset_batch_line_views.xml",
