+++ purchase_contract_kmitl/__manifest__.py
@@ -5,7 +5,7 @@
     "author": "Aginix Technologies",
     "website": "https://github.com/aginix/kmitl",
     "category": "KMITL",
-    'depends': ['hr_department_short_name', 'purchase_exception', 'purchase_order_department', 'purchase_invoice_plan', 'purchase_budget'],
+    'depends': ['hr_department_short_name', 'purchase_exception', 'purchase_order_kmitl', 'purchase_invoice_plan', 'purchase_budget'],
     "data": [
         "data/cron.xml",
         "data/purchase_contract_type.xml",
