--- purchase_order_department/__manifest__.py
@@ -1,16 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    'name': 'Purchase Order Department',
-    'version': '16.0.1.0.0',
-    'summary': """ Purchase Order Department Summary """,
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "category": "KMITL",
-    'depends': ['purchase', 'hr'],
-    "data": [
-        "views/purchase_order_views.xml"
-    ],
-    'installable': True,
-    'auto_install': False,
-    'license': 'LGPL-3',
-}
