--- purchase_order_department/i18n/th.po
@@ -1,28 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* purchase_order_department
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-11-26 09:24+0000\n"
-"PO-Revision-Date: 2025-11-26 16:24+0700\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"Language: th\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
-"X-Generator: Poedit 3.6\n"
-
-#. module: purchase_order_department
-#: model:ir.model.fields,field_description:purchase_order_department.field_purchase_order__department_id
-msgid "Department"
-msgstr "ส่วนงาน"
-
-#. module: purchase_order_department
-#: model:ir.model,name:purchase_order_department.model_purchase_order
-msgid "Purchase Order"
-msgstr "คำสั่งซื้อ"
