--- purchase_order_department/models/purchase_order.py
@@ -1,24 +1,0 @@
-# -*- coding: utf-8 -*-
-import logging
-
-from odoo import models, fields, api, _
-from odoo.exceptions import UserError, ValidationError
-
-_logger = logging.getLogger(__name__)
-
-
-class PurchaseOrder(models.Model):
-    _inherit = 'purchase.order'
-
-    READONLY_STATES = {
-        'purchase': [('readonly', True)],
-        'done': [('readonly', True)],
-        'cancel': [('readonly', True)],
-    }
-
-    department_id = fields.Many2one(
-        comodel_name="hr.department",
-        string="Department",
-        states=READONLY_STATES,
-        tracking=True
-    )
