--- purchase_order_department/views/purchase_order_views.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-  <odoo>
-      
-    <!-- View purchase.order form -->
-    <record id="view_purchase_order_form" model="ir.ui.view">
-        <field name="name">view.purchase.order.form</field>
-        <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase.purchase_order_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='company_id']" position="after">
-              <field name="department_id" />
-            </xpath>
-        </field>
-    </record>
-
-  </odoo>
