--- purchase_order_department_link_by_purchase_request/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import wizards
