--- purchase_order_department_link_by_purchase_request/__manifest__.py
@@ -1,15 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    'name': 'Purchase Order Department Link By Purchase Request',
-    'version': '16.0.1.0.0',
-    'summary': """ Purchase Order Department Link By Purchase Request Summary """,
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "category": "KMITL",
-    'depends': ['purchase_request_department', 'purchase_order_department'],
-    "data": [
-    ],
-    'installable': True,
-    'auto_install': False,
-    'license': 'LGPL-3',
-}
