--- purchase_order_department_operating_unit/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import models
