--- purchase_order_department_operating_unit/views/purchase_order_views.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-  <odoo>
-      
-    <!-- View purchase.order form -->
-    <record id="view_purchase_ou_order_form" model="ir.ui.view">
-        <field name="name">view.purchase.ou.order.form</field>
-        <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase_operating_unit.purchase_order_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='operating_unit_id']" position="attributes">
-              <attribute name="invisible">1</attribute>
-            </xpath>
-            <xpath expr="//field[@name='department_id']" position="attributes">
-              <attribute name="domain">['|', ('operating_unit_id', '=', False), ('operating_unit_id', '=', operating_unit_id)]</attribute>
-            </xpath>
-
-            <xpath expr="//group[@name='other_info']" position="inside">
-              <field name="operating_unit_id" groups="operating_unit.group_multi_operating_unit"/>
-            </xpath>
-        </field>
-    </record>
-
-  </odoo>
