+++ purchase_order_kmitl/__manifest__.py
@@ -5,7 +5,16 @@
     "author": "Aginix Technologies",
     "website": "https://github.com/aginix/kmitl",
     "category": "KMITL",
-    "depends": ["purchase", "purchase_requisition", "purchase_no_rfq", "account_fiscal_year"],
+    "depends": [
+        "purchase",
+        "purchase_requisition",
+        "purchase_no_rfq",
+        "account_fiscal_year",
+        "hr",
+        "purchase_operating_unit",
+        "hr_department_operating_unit",
+        "purchase_request_department",
+    ],
     "data": ["views/purchase_order_views.xml"],
     "application": False,
     "installable": True,
