+++ purchase_order_kmitl/models/purchase_order.py
@@ -26,3 +26,32 @@ class PurchaseOrder(models.Model):
     date_planned = fields.Datetime(
         string="Date End"
     )
+
+    # From purchase_order_department
+    department_id = fields.Many2one(
+        comodel_name="hr.department",
+        string="Department",
+        states=READONLY_STATES,
+        tracking=True
+    )
+
+    # From purchase_order_department_operating_unit
+    operating_unit_id = fields.Many2one(
+        compute="_compute_operating_unit_id",
+        default=lambda self: self._default_operating_unit_id(),
+        store=True,
+    )
+
+    def _default_operating_unit_id(self):
+        department_id = self.env.user.employee_id.department_id
+        if department_id and department_id.operating_unit_id:
+            return department_id.operating_unit_id.id
+        return self.env["res.users"].operating_unit_default_get()
+
+    @api.depends("department_id")
+    def _compute_operating_unit_id(self):
+        for rec in self:
+            if rec.department_id and rec.department_id.operating_unit_id:
+                rec.operating_unit_id = rec.department_id.operating_unit_id.id
+            else:
+                rec.operating_unit_id = False
