+++ purchase_order_kmitl/wizards/purchase_request_line_make_purchase_order.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
-
 import logging
+
 from odoo import models, fields, api, _
-from odoo.exceptions import UserError, ValidationError
 
 _logger = logging.getLogger(__name__)
