+++ purchase_order_ux/__manifest__.py
@@ -6,7 +6,7 @@
     "author": "Aginix Technologies",
     "website": "https://github.com/aginix/kmitl",
     "category": "KMITL",
-    'depends': ['purchase_order_kmitl', 'purchase_contract_kmitl', 'purchase_invoice_plan' ,'purchase_order_department' ,'purchase_order_payment_type'],
+    'depends': ['purchase_order_kmitl', 'purchase_contract_kmitl', 'purchase_invoice_plan', 'purchase_order_payment_type'],
     "data": [
         "views/purchase_order_views.xml"
     ],
